Birdenst generator updates
