+++ birdnest_generator/model.py
@@ -53,55 +53,3 @@ class Report_wrap(BaseModel):
     report: Report
 
 
-
-
-if __name__=='__main__':
-    # Generate drones and pilots
-    all_drones = []
-    for i in range(20):
-        all_drones.append(Drone())
-
-    pilots = []
-    for i in range(10):
-        pilots.append(Pilot(drones=all_drones[i*2:i*2+2]))
-
-    ic(pilots)
-
-    # Generate report
-    loop = True
-    while loop == True:
-        t0 = time.time()
-
-        drones = all_drones.copy()
-        for i, drone in enumerate(drones):
-            drones[i].positionY = random.randint(-700000, 700000)
-            drones[i].positionX = random.randint(-700000, 700000)
-            drones[i].altitude=random.randint(-700000, 700000)
-
-        for i, drone in enumerate(drones):
-            if drone.positionX not in range(-500000, 500000) or drone.positionY not in range(-500000, 500000):
-                drones.pop(i)
-
-        updateIntervalMs = 2000
-        report = Report(
-            deviceInformation=DeviceInformation(listenRange=500000, deviceStarted='2022-12-23T19:59:46.911Z', updateIntervalMs=2000),
-            capture=(Capture(drone=drones))
-        )
-        # Make Report wrap becouse xmltodict required one root element
-        report = Report_wrap(report=report)
-
-        # Make dict
-        report_dict = report.model_dump()
-
-        # Change name deviceId to @deviceId to send it as tag parameter
-        report_dict['report']['deviceInformation']['@deviceId'] = report_dict['report'].pop('deviceId')
-        report_dict['report']['capture']['@snapshotTimestamp'] = report_dict['report'].pop('snapshotTimestamp')
-
-        print(xmltodict.unparse(report_dict, pretty=True))
-
-        t1 = time.time()
-        code_speed = t1-t0
-        ic(code_speed)
-
-        loop = False
-        time.sleep(2)
